Change is_ssl() function to work with containers
This is a hack to make the Wordpress more container friendly for TLS.  Right
now it tries to guess the TLS scheme by introspecting whether the server
it's running on is using TLS or not, and it ignores the 'siteurl'.  The
problem with this is that if you want to do TLS termination by bolting on
a proxy container, it will munge many of your URLs to be "http://" instead
of "https://".

This change makes it so that the is_ssl() function checks the 'siteurl'
scheme instead.