Redirect to HTTPS without a trailing slash
The mod_dir redirects https://example.com/wp-admin to http://example.com/wp-admin/ when the apache is behind an HTTPS reverse proxy because the Location header requires the scheme and the apache didn't consider X-Forwarded-Proto.

This rewrite rule in this htaccess will add a trailing slash with the correct scheme.

Ref: https://serverfault.com/a/851918
Ref: https://stackoverflow.com/a/40447815

%{LA-U:REQUEST_FILENAME} will access the server via mod_dir and check if the resulted URL is a directory.